change: move contextual helper to composable
